# Conflicts:
#	app_pojavlauncher/src/main/res/layout-land/fragment_launcher.xml
#	app_pojavlauncher/src/main/res/layout/fragment_launcher.xml

+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/EfficientAndroidLWJGLKeycode.java
@@ -198,7 +198,7 @@ public class EfficientAndroidLWJGLKeycode {
         sendKeyPress(getValueByIndex(index));
     }
 
-    public static int getValueByIndex(int index) {
+    public static short getValueByIndex(int index) {
         return sLwjglKeycodes[index];
     }
